refactor: extract analyzer orchestrator tests into tests.rs (#190)
Moves the inline #[cfg(test)] mod tests block out of analyzer/mod.rs
into a dedicated tests.rs file (wired via `mod tests;`). Per-concern
tests live alongside their submodules (e.g. coupling::tests, layers
test mod via inherent impls, etc.); tests.rs covers integration-style
tests that exercise audit() and audit_with_settings() end-to-end.
